generation/requirements.txt
click==8.1.7
pyairtable==2.3.3
python-dotenv==1.0.1
python_igraph==0.11.6
Requests==2.32.3
spacy==3.7.6
